requirements.txt
python-telegram-bot==20.8
Flask==2.3.3
requests==2.31.0